+++ api/src/main/java/net/jqwik/api/SampleReportingFormat.java
@@ -6,7 +6,7 @@ import org.apiguardian.api.*;
 
 import static org.apiguardian.api.API.Status.*;
 
-@API(status = EXPERIMENTAL, since = "1.3.1")
+@API(status = MAINTAINED, since = "1.4.0")
 public interface SampleReportingFormat extends Comparable<SampleReportingFormat> {
 
 	/**