+++ api/src/main/java/net/jqwik/api/arbitraries/FloatArbitrary.java
@@ -2,8 +2,6 @@ package net.jqwik.api.arbitraries;
 
 import org.apiguardian.api.*;
 
-import net.jqwik.api.*;
-
 import static org.apiguardian.api.API.Status.*;
 
 /**
@@ -71,6 +69,6 @@ public interface FloatArbitrary extends NumericalArbitrary<Float, FloatArbitrary
 	/**
 	 * Set shrinking target to {@code target} which must be between the allowed bounds.
 	 */
-	@API(status = EXPERIMENTAL, since = "1.1.5")
+	@API(status = MAINTAINED, since = "1.4.0")
 	FloatArbitrary shrinkTowards(float target);
 }