+++ api/src/main/java/net/jqwik/api/arbitraries/LongArbitrary.java
@@ -2,8 +2,6 @@ package net.jqwik.api.arbitraries;
 
 import org.apiguardian.api.*;
 
-import net.jqwik.api.*;
-
 import static org.apiguardian.api.API.Status.*;
 
 /**
@@ -32,6 +30,6 @@ public interface LongArbitrary extends NumericalArbitrary<Long, LongArbitrary> {
 	/**
 	 * Set shrinking target to {@code target} which must be between the allowed bounds.
 	 */
-	@API(status = EXPERIMENTAL, since = "1.1.4")
+	@API(status = MAINTAINED, since = "1.4.0")
 	LongArbitrary shrinkTowards(long target);
 }